+++ shell/platform/android/io/flutter/embedding/android/FlutterActivityAndFragmentDelegate.java
@@ -631,7 +631,7 @@ import java.util.Arrays;
     if (flutterEngine != null) {
       // This is always an indication that the Dart VM should collect memory
       // and free any unneeded resources.
-      flutterEngine.notifyLowMemoryWarning();
+      flutterEngine.getDartExecutor().notifyLowMemoryWarning();
       // Use a trim level delivered while the application is running so the
       // framework has a chance to react to the notification.
       if (level == TRIM_MEMORY_RUNNING_LOW) {
@@ -654,7 +654,7 @@ import java.util.Arrays;
   void onLowMemory() {
     Log.v(TAG, "Forwarding onLowMemory() to FlutterEngine.");
     ensureAlive();
-    flutterEngine.notifyLowMemoryWarning();
+    flutterEngine.getDartExecutor().notifyLowMemoryWarning();
     flutterEngine.getSystemChannel().sendMemoryPressureWarning();
   }
