+++ shell/platform/android/io/flutter/embedding/engine/FlutterJNI.java
@@ -824,12 +824,11 @@ public class FlutterJNI {
   private native Bitmap nativeGetBitmap(long nativePlatformViewId);
 
   /**
-   * Notifies the Dart VM of a low memory event, or that the application is in
-   * a state such that now is an appropriate time to free resources, such as
-   * going to the background.
+   * Notifies the Dart VM of a low memory event, or that the application is in a state such that now
+   * is an appropriate time to free resources, such as going to the background.
    *
-   * This is distinct from sending a SystemChannel message about low memory,
-   * which only notifies the running Flutter application.
+   * <p>This is distinct from sending a SystemChannel message about low memory, which only notifies
+   * the running Flutter application.
    */
   @UiThread
   public void notifyLowMemoryWarning() {
